refactor: move Lotto.js to model directory
